requirement.txt
bitarray==2.6.2
bitstring==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
cuckoo-filter==1.0.6
Flask==2.2.3
Flask-FontAwesome==0.1.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mmh3==3.0.0
mysql-connector-python==8.0.32
numpy==1.23.5
protobuf==3.20.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.16.0
pyOpenSSL==23.0.0
rsa==4.9
Werkzeug==2.2.3
